Add abstraction for hyperparameter marginalization
Hyperparameter marginalization is really its own tasks, and there isn't
just "one" way to do it.

Abstracting away from it lets users implement their own algorithms for
hyperparameter marginalization in the future, should they want it.
For example a very cheap (but rough) approach would be to simply form a
Laplace approximation around the mode and then marginalize from that.
This is now quite easy to implement without making deep interface changes.

It's also just a nice separation of concerns.